requirements.txt
Django==4.2.5
channels[daphne]
channels-redis==4.1.0
django-autoslug==1.9.9
psycopg2-binary==2.9.7